frontend/requirements-frontend.txt
streamlit==1.26.0
requests>=2.31.0
pillow>=10.0.0

